refactor: DropModal
